Monadify setup
