+++ core-strato/strato-init/exec_src/Main.hs
@@ -3,6 +3,7 @@
 import           HFlags
 
 import           Blockchain.Setup
+import           Blockchain.Init.Options()
 import           Blockchain.Output() -- Import the --minLogLevel flag
 
 defineFlag "genesisBlockName" "livenet" "use the alternate stablenet genesis block"
